refactor: deprecated된 메서드 제거
