+++ NearbyNetwork/NearbyNetworkTests/NearbyNetworkTests.swift
@@ -11,14 +11,14 @@ import Network
 import XCTest
 
 final class NearbyNetworkTests: XCTestCase {
-    var nearbyNetworkService: RefactoredNearbyNetworkService?
+    var nearbyNetworkService: NearbyNetworkService?
     var mockBrowser: NWBrowser?
     var mockListener: NWListener?
     let serviceName = "airplain"
     let serviceType = "_airplain._tcp"
 
     override func setUpWithError() throws {
-        nearbyNetworkService = RefactoredNearbyNetworkService(
+        nearbyNetworkService = NearbyNetworkService(
             myPeerID: UUID(),
             serviceName: serviceName,
             serviceType: serviceType)
@@ -100,7 +100,7 @@ final class NearbyNetworkTests: XCTestCase {
         let advertisingExpectation = XCTestExpectation(description: "광고 성공 여부")
         var advertisingHostName: String?
         var advertisingPeerInfo: [String]?
-        let foundPeerHandler: ([RefactoredNetworkConnection]) -> Void = { connection in
+        let foundPeerHandler: ([NetworkConnection]) -> Void = { connection in
             advertisingHostName = connection.first!.name
             advertisingPeerInfo = connection.first!.connectedPeerInfo
             advertisingExpectation.fulfill()
